+++ src/main/java/org/jallen/tyrael/entity/Base.java
@@ -1,0 +1,6 @@
+package org.jallen.tyrael.entity;
+
+@FunctionalInterface
+public interface Base {
+  public String showMessage(String s); 
+}