+++ src/main/java/org/jallen/tyrael/entity/Skills.java
@@ -1,0 +1,70 @@
+package org.jallen.tyrael.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+
+@Entity
+@Table(name = "skills")
+public class Skills {
+
+  @Id
+  @Column(name = "id", nullable = false)
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
+
+  @Column(columnDefinition = "TEXT")
+  private String name;
+
+  @Column(columnDefinition = "TEXT")
+  private Long experience;
+
+  @Column(columnDefinition = "TEXT")
+  private String description;
+
+  @Column(columnDefinition = "TEXT")
+  private String type;
+
+  public Long getId() {
+    return id;
+  }
+
+  public void setId(Long id) {
+    this.id = id;
+  }
+
+  public String getName() {
+    return name;
+  }
+
+  public void setName(String name) {
+    this.name = name;
+  }
+
+  public Long getExperience() {
+    return experience;
+  }
+
+  public void setExperience(Long experience) {
+    this.experience = experience;
+  }
+
+  public String getDescription() {
+    return description;
+  }
+
+  public void setDescription(String description) {
+    this.description = description;
+  }
+
+  public String getType() {
+    return type;
+  }
+
+  public void setType(String type) {
+    this.type = type;
+  }
+}