+++ src/main/java/org/jallen/tyrael/mapper/ApplicationMapper.java
@@ -1,5 +1,0 @@
-package org.jallen.tyrael.mapper;
-
-public class ApplicationMapper {
-
-}