+++ src/main/java/org/jallen/tyrael/repositories/SkillsRepository.java
@@ -1,8 +1,8 @@
 package org.jallen.tyrael.repositories;
 
-import org.jallen.tyrael.entity.Application;
+import org.jallen.tyrael.entity.Skills;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ApplicationRepository extends JpaRepository<Application, Long> {}
+public interface SkillsRepository extends JpaRepository<Skills, Long> {}