+++ CHANGELOG.md
@@ -2,6 +2,35 @@
 
 All notable changes to this project will be documented in this file.
 
+# 1.2.0
+
+# 1.2.0 Features
+
+- `[AboutController]` Added controller for About entity.
+- `[ExperienceController]` Added controller for Experience entity.
+- `[InformationController]` Added controller for Information entity.
+- `[SkillsController]` Added controller for Skills entity.
+- `[About]` Added About entity.
+- `[Experience]` Added Experiences entity.
+- `[Information]` Added Information entity.
+- `[Skills]` Added Skills entity.
+- `[ExperienceMapper]` Added experience mapper for mapping purposes.
+- `[AboutRepository]` Added repository for About entity.
+- `[ExperienceRepository]` Added repository for Experience entity.
+- `[InformationRepository]` Added repository for Information entity.
+- `[SkillsRepository]` Added repository for Skills entity.
+- `[AboutService]` Added service for About entity.
+- `[ExperienceService]` Added controller for Experience entity.
+- `[InformationService]` Added service for Information entity.
+- `[SkillsService]` Added service for Skills entity.
+
+# 1.2.0 Fixes
+
+- `[Controller]` Code cleanup for all controllers.
+- `[Entity]` Code cleanup for all entities.
+- `[Service]` Code cleanup for all services.
+- `[Repository]` Code cleanup for all repositories.
+
 # 1.1.0
 
 ## 1.1.0 Features
@@ -9,9 +38,6 @@ All notable changes to this project will be documented in this file.
 - `[ApplicationController]` Added web service /pages.
 - `[ApplicationMapper]` Added mapper for Application.
 - `[ReviewController]` Added controller for review purposes.
-- `[Animal]` Added animal entity for review purposes.
-- `[Dog]` Added dog entity for review purposes.
-- `[Employee]` Added dog entity for review purposes.
 
 # 1.0.0